A few more modifications since require seems to work with webpack
