+++ src/App.js
@@ -1,6 +1,7 @@
 import React from "react";
 //import main from "./index.html";
 import "./App.css";
+import * as apis from "./scripts/apis";
 
 function App() {
   return <div>This needs to be fixed</div>;
